refactor: EngineViewModel.cs - separate mods logic to ModService.cs - disable interaction Pandora Base Mod in UI (always enable) - refactoring behaviors with mods
