+++ Pandora Behaviour Engine/Assets/CutsomStyles.axaml
@@ -1,7 +1,5 @@
 <Styles xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
-			  xmlns:idd="clr-namespace:Avalonia.Xaml.Interactions.DragAndDrop;assembly=Avalonia.Xaml.Interactions.DragAndDrop"
         xmlns:b="using:Pandora.Behaviors"
         xmlns:ui="using:FluentAvalonia.UI.Controls">
   <Design.PreviewWith>
@@ -17,12 +15,12 @@
   <!-- Add Styles Here -->
   <Style Selector="DataGrid.DragAndDrop">
     <Setter Property="RowHeaderWidth" Value="28" />
-    <Setter Property="(i:Interaction.Behaviors)">
-      <i:BehaviorCollectionTemplate>
-        <i:BehaviorCollection>
-          <idd:ContextDropBehavior Handler="{StaticResource ModsDataGridDropHandler}" />
-        </i:BehaviorCollection>
-      </i:BehaviorCollectionTemplate>
+    <Setter Property="(Interaction.Behaviors)">
+      <BehaviorCollectionTemplate>
+        <BehaviorCollection>
+          <ContextDropBehavior Handler="{StaticResource ModsDataGridDropHandler}" />
+        </BehaviorCollection>
+      </BehaviorCollectionTemplate>
     </Setter>
   </Style>
 
@@ -42,12 +40,12 @@
     </Setter>
   </Style>
   <Style Selector="DataGrid.DragAndDrop DataGridRowHeader">
-    <Setter Property="(i:Interaction.Behaviors)">
-      <i:BehaviorCollectionTemplate>
-        <i:BehaviorCollection>
-          <b:ContextDragWithDirectionBehavior HorizontalDragThreshold="3" VerticalDragThreshold="3" />
-        </i:BehaviorCollection>
-      </i:BehaviorCollectionTemplate>
+    <Setter Property="(Interaction.Behaviors)">
+      <BehaviorCollectionTemplate>
+        <BehaviorCollection>
+          <ContextDragBehavior HorizontalDragThreshold="3" VerticalDragThreshold="3" />
+        </BehaviorCollection>
+      </BehaviorCollectionTemplate>
     </Setter>
     <Setter Property="Content">
       <Template>
@@ -59,12 +57,12 @@
   </Style>
 
   <Style Selector="DataGrid.ItemsDragAndDrop">
-    <Setter Property="(i:Interaction.Behaviors)">
-      <i:BehaviorCollectionTemplate>
-        <i:BehaviorCollection>
-          <idd:ContextDropBehavior Handler="{StaticResource ModsDataGridDropHandler}" />
-        </i:BehaviorCollection>
-      </i:BehaviorCollectionTemplate>
+    <Setter Property="(Interaction.Behaviors)">
+      <BehaviorCollectionTemplate>
+        <BehaviorCollection>
+          <ContextDropBehavior Handler="{StaticResource ModsDataGridDropHandler}" />
+        </BehaviorCollection>
+      </BehaviorCollectionTemplate>
     </Setter>
   </Style>
