--- Pandora Behaviour Engine/Command/RelayCommand.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Windows.Input;
-
-namespace Pandora.Command
-{
-    public class RelayCommand : ICommand
-    {
-        private readonly Action<object?> execute;
-        private readonly Func<object?, bool>? canExecute;
-
-        public RelayCommand(Action<object?> execute, Func<object?, bool>? canexecute = null)
-        {
-            ArgumentNullException.ThrowIfNull(execute);
-            this.execute = execute;
-            this.canExecute = canexecute;
-        }
-        public event EventHandler? CanExecuteChanged;
-        public void RaiseCanExecuteChanged() => CanExecuteChanged?.Invoke(this, EventArgs.Empty);
-        public bool CanExecute(object? parameter) => canExecute is null || canExecute(parameter);
-        public void Execute(object? parameter) => execute(parameter);
-    }
-}
