+++ Pandora Behaviour Engine/Pandora Behaviour Engine.csproj
@@ -26,6 +26,7 @@
 	  <PackageReference Condition="'$(Configuration)' == 'Release'" Include="Avalonia.Controls.DataGrid" Version="11.3.0" />
     <PackageReference Condition="'$(Configuration)' == 'Debug'" Include="Avalonia.Diagnostics" Version="11.3.0" />
     <PackageReference Include="Avalonia.ReactiveUI" Version="11.3.0" />
+    <PackageReference Include="Avalonia.Xaml.Behaviors" Version="11.3.0" />
     <PackageReference Include="Avalonia.Xaml.Interactions.DragAndDrop" Version="11.3.0" />
     <PackageReference Include="DynamicData" Version="9.2.2" />
     <PackageReference Include="FluentAvaloniaUI" Version="2.3.0" />
